backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Django==5.2
django-cors-headers==4.7.0
django-decouple==2.1
django-filter==25.1
django-leaflet==0.31.0
django-redis==5.4.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
executing==2.2.0
fiona==1.10.1
folium==0.19.6
fonttools==4.58.0
GeoAlchemy2==0.17.1
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
geoserver-rest==2.9.1
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==5.4.0
mapclassify==2.8.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.15.3
shapely==2.1.1
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
xmltodict==0.14.2
xyzservices==2025.4.0
